+++ OneSila/core/managers/__init__.py
@@ -1,0 +1,24 @@
+from .multi_tenant import MultiTenantUserLoginTokenQuerySet, \
+    MultiTenantUserLoginTokenManager, MultiTenantQuerySet, \
+    MultiTenantManager, MultiTenantUserManager, MultiTenantCompanyCreateMixin, \
+    QuerySetProxyModelMixin
+from .shared import QuerySet, Manager
+
+# from django.db.models import QuerySet as DjangoQueryset
+# from django.db.models import Manager as DjangoManager
+# from django.db import IntegrityError
+# from django.core.exceptions import ValidationError
+# from django.db import models
+# from django.db.models.constants import LOOKUP_SEP
+# from django.core.exceptions import FieldDoesNotExist
+# from django.utils.text import smart_split, unescape_string_literal
+# from django.contrib.auth.models import BaseUserManager
+# from django.contrib.admin.utils import lookup_spawns_duplicates as lookup_function
+
+# from core.exceptions import SearchFailedError
+
+# import operator
+# from functools import reduce
+
+# import logging
+# logger = logging.getLogger(__name__)
