+++ OneSila/core/schema/core/types/filters.py
@@ -7,6 +7,7 @@ class SearchFilterMixin:
 
     def filter_search(self, queryset):
         if self.search not in (None, UNSET):
+            print(set(queryset.values_list('multi_tenant_company', flat=True)))
             queryset = queryset.search(self.search)
 
         return queryset
